Single card start
