+++ frontend/src/components/CardInfo.tsx
@@ -1,0 +1,3 @@
+const CardInfo = () => {
+  return;
+};
