+++ frontend/src/components/CardLegality.tsx
@@ -1,0 +1,3 @@
+const CardLegality = () => {
+  return;
+};
