+++ frontend/src/components/CardMarket.tsx
@@ -1,0 +1,3 @@
+const CardMarket = () => {
+  return;
+};
