+++ frontend/src/main.tsx
@@ -1,10 +1,19 @@
-import React from 'react';
+import {
+  createBrowserRouter,
+  Link,
+  RouterProvider,
+  Outlet,
+} from 'react-router-dom';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
 
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <App></App>,
+  },
+]);
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <RouterProvider router={router}></RouterProvider>
 );
